# express-api/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Simple CRUD TODO API",
      "description": "This expressJS application demos simple todo GET, PUT & DELETE functionalities with swagger UI",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "tags": [
      {
        "name": "Todo",
        "description": "API for Todos in the system"
      }
    ],
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
      "/todo":{
          "get": {
            "tags": [
              "Todo"
            ],
            "summary": "Get all todos",
            "responses": {
              "200": {
                "description": "OK",
                "schema": {
                  "$ref": "#/definitions/Response"
                }
              }
            }
          },
          "post": {
            "summary": "Add todo with given name",
            "tags": [
              "Todo"
            ],
            "description": "Create new todo in system",
            "parameters": [
              {
                "name": "todo",
                "in": "body",
                "description": "Todo name that we want to create",
                "schema": {
                  "$ref": "#/definitions/TodoCreate"
                }
              }
            ],
            "produces": [
              "application/json"
            ],
            "responses": {
              "200": {
                "description": "New todo is created",
                "schema": {
                  "$ref": "#/definitions/Response"
                }
              }
            }
          },
          "delete": {
            "summary": "Delete todo with given name & date",
            "tags": [
              "Todo"
            ],
            "parameters": [
                {
                    "name": "todo",
                    "in": "body",
                    "description": "Todo name that we want to delete",
                    "schema": {
                      "$ref": "#/definitions/Todo"
                    }
                  }
            ],
            "produces": [
                "application/json"
            ],
            "responses": {
              "200": {
                "description": "Todo is deleted",
                "schema": {
                  "$ref": "#/definitions/Response"
                }
              }
            }
          }
      }
    },
    "definitions": {
          "Todo": {
            "required": [
              "name",
              "addedDate"
            ],
            "properties": {
              "name": {
                "type": "string"
              },
              "addedDate": {
                "type": "string"
              }
            }
          },
          "TodoCreate": {
            "required": [
              "todo"
            ],
            "properties": {
              "todo": {
                "type": "string"
              }
            }
          },
          "Response": {
            "properties": {
                "success": {
                    "type": "boolean"
                  },
                  "data": {
                    "type": "string"
                  }
                }
          }
    }
  }